requirements-prd.txt
pynie==0.0.2
loguru
minio